app_name/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bitstring==3.1.9
blinker==1.4
certifi==2021.10.8
click==8.0.3
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
esptool==3.2
Flask==2.0.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
flask-restx==0.5.1
Flask-WTF==1.0.0
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
mongoengine==0.23.1
pymongo==3.12.3
pyrsistent==0.18.1
pyserial==3.5
python-decouple==3.6
pytz==2021.3
reedsolo==1.5.4
six==1.16.0
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
